fix: use != instead of is not (#13)
Based on https://realpython.com/python-is-identity-vs-equality/, `is not` operator compares if two variables refer to the same object in memory. Hence switching to != and == comparisons. Although the types being compared are `int` and fall within range, it seems like the comaprison being made is incorrect.

Why only the bug reporter can reproduce this but not any of us is still a mystery to me. So unfortunately, I cannot test this fix apart from some theoretical reading that I did.

Nonetheless, this should fix it for everyone is my solid hope.

```
>>> import momento_wire_types.cacheclient_pb2 as cache_client_types
>>> get_resp = cache_client_types.GetResponse()
>>> type(get_resp)
<class 'cacheclient_pb2.GetResponse'>
>>> type(get_resp.result)
<class 'int'>
>>> type(cache_client_types.Hit)
<class 'int'>
```