feat: add asyncio client (#28)
* feat: add asyncio client

In a way similar to grpc, this adds an `aio` module. Over time
we should consider subsuming the synchronous client under the
asynchronous implementation (easier to make async sync than
the other way around).

I simplified the whole interceptor thing to make fewer classes
and to keep the channel stuff easier to follow. I'm not sure
the manager is adding a lot atm but I'm leaving that much in
case it's useful for multi-cache-client-wrapper stuff.

Also added type hints all over the place in the aio stuff. I
think we should _always_ use type hints.

Cloned the test_momento file; async calling convention is
different, so we don't have a convenient means of combining
these. And since the asynchronous client can be used in a
synchronous context we should be replacing the synchronous
client with a thin proxy layer to the async client. When
we do so we can delete the test_momento.py copy.

Also fixes the README command for tests. Idk how to make the
original form work but unittest discover works fine. Possibly
related, I also needed to fix the import path for these tests.
I weakly suspect that the previous author's `pythonpath` had
`src` or `src/momento` directly added to it. I updated ci.yml
with this discover command as well.