fix: Handle GRPC Error Codes. (#42)
* fix: Rename Error and tighten cache name validations

Fix for #23 - CacheValueError is a confusing name for surfacing INVALID_ARGUMENT errors. It can be easily confused with something related to setting a value in Cache. Proposing `CacheValueError` is renamed to `InvalidArgumentError`

Fix for #24 - Given the type-safety in python, customers can accidentally pass unsupported types. This results in a ClientSdkException when in fact it is an Invalid Input Error. We already type check our keys, values, ttl, so just adding one for cache name. Though it may be just wise to let the type errors pass through.

* fix broken error

* fix tests

* rename errors

* remove unwanted import

* update check

* fix: use None check, not seems to checking empty str that we don't want to check on SDK

* fix: doc strings

* Update src/momento/errors.py

Co-authored-by: Ruth Linehan <1530016+rlinehan@users.noreply.github.com>

* fix docstrings

* rename exception

* update docstring

* Update src/momento/simple_cache_client.py

Co-authored-by: Ruth Linehan <1530016+rlinehan@users.noreply.github.com>

* fix error message

Co-authored-by: Ruth Linehan <1530016+rlinehan@users.noreply.github.com>