fix: correct some issues with CI/CD jobs, move build to `tox` (#49)
* fix: correct some issues with CI/CD jobs, move build to `tox`

This commit introduces `tox` for our builds.  It replaces the
previous approach of directly creating a virtualenv and defers
that to `tox`, which will create multiple virtualenvs to allow
testing of multiple versions of python.

`tox` is used in a wide variety of high-visibility python projects
such as pytest, Django Channels, Twisted, etc.  In addition to
allowing us to more easily test multiple python versions, it
sounds like it will actual install the built package into the
virtualenvs and then run the tests against that, rather than
running against the source, so it should give us better coverage
of the packaging itself.

It also gives us an easy pattern for adding linting and some other
tasks in subsequent commits.

Also in this commit, we make some changes to the CI/CD github
actions:
1. Previously the tests were only being run in the CI job, which,
   on push, was running in parallel to the CD job.  This meant that
   it was possible for the tests to fail but we would still publish
   the new artifact.  This is now fixed, the CD job also runs the
   tests.
2. Build commands switched to tox
3. CD job no longer runs on push; we will manually run it to do
   releases on a more deliberate schedule for now.