chore: temporarily remove python 3.6
In our async tests we are using a construct that isn't available
until python 3.8, so the tests fail if we try to run them on 3.6.

We need to fix this but it will be better to do it in a separate
PR, so this commit temporarily removes python 3.6 from the test
matrix.