chore: add mypy type checker to build commands (#51)
* chore: add mypy type checker to build commands

This commit adds a `lint` target to the tox build environments,
and runs it in CI.

As of this commit the `lint` target just runs `mypy`, which is
a type checker that ensures that all methods have type annotations
and that they are correct.  (We should be using this on all of our
python projects going forward.)

Also in this commit:

* Add an initial mypy config and fix all of the type hinting
  errors that mypy found in the project.
* Add python 3.6 to the test matrix since our docs claim that we
  are compatible with 3.6 or later.

* chore: temporarily remove python 3.6

In our async tests we are using a construct that isn't available
until python 3.8, so the tests fail if we try to run them on 3.6.

We need to fix this but it will be better to do it in a separate
PR, so this commit temporarily removes python 3.6 from the test
matrix.