chore: add 'black' formatter to tox build environment (#52)
* chore: add 'black' formatter to tox build environment

This commit adds the 'black' formatter to our 'lint' build
environment.  We run it in "check" mode, to fail builds if
the code isn't formatted properly.

We also add a "format" target to run black in the mode where
it will actually fix the formatting, for use by local devs
(via `tox -e format`).

This commit also includes the results of applying black to
the existing codebase to standardize the existing formatting.