fix: mitigate m1 incompatibility
There is a bug in grpc's packaging of native artifacts that relates
to certain versions of python 3 and the default wheel tags.
Investigation is ongoing, but in the meantime there is a simple
mitigation that users can try to get around the issue for local
development workflows.

Relates to: https://github.com/grpc/grpc/issues/28387