feat: remove incubating, get-multi, and set-multi client-side apis (#156)
This repo is where we did early prototyping of many APIs to get
a feel for the UX for them before we finalized the wire format
and completed the server-side implementations.

As we ramp up for the 1.0 release of the Python client, we will
be adding in the final versions of these APIs that correspond
to the implementations that are now available on the server.

In preparation for that, and to accomodate some forthcoming
refactors that we need to do to improve asyncio compatibility
and error handling UX, this commit removes all of the client-side
implementations for get-multi, set-multi, and collection data
structures. We will add them back in soon, after the error handling
changes have landed.