fix: re-introduce retry interceptor for synchronous client
In an earlier commit we made some changes to the implementation
of the synchronous client to use the grpc synchronous API directly,
rather than wrapping the asyncio API, which was causing problems
in environments that used other asyncio libraries.

During the rollout of that change we mistakenly lost the retry
capabilities that we have in the asyncio client.  This commit
adds the retry suppport back in for the sync client.

This implementation is not DRY and we need to clean it up as we
are doing the work to introduce Configuration, but this will
at least unblock users for now.