chore: pin cryptography dependency
We need the crytography dependency for pyjwt. However the pyjwt crypto
extra uses version >=3.4, which is unconstrained.

Since the crytography package depends on glibc, which is external to
python, the crytography version introduces incompatibilities in
specific environments (eg Lambda). We pin to 3.4.8 to prevent a moving
target.

see: https://github.com/jpadilla/pyjwt/issues/800