feat: break out examples by python version and function; use poetry
To better organize the code and improve compatibility with dependabot,
we separate the examples by python version and load gen.

The first reason for this is the different examples have different
dependencies. The load generator a depends on HDR Histogram while the
simple examples do not. Customers have complained about why they have
to install HDR Histogram to run a specific example.

A second reason is dependabot expects a directory to run the updated
code in. If the directory contains version-incompatible code (we
cannot run the py310 code using py37) then dependabot will fail. Hence
we need separate directories for those.

Lastly we standardize the use of poetry. Before we use poetry for the
main project but pipenv for the examples.