chore: undo public type aliases
We use concrete type aliases instead of type aliases for the cache
client and responses. We do this because neither VSCode nor IntelliJ
resolve the type hints in methods when consuming the library. Ie when
typing `client.dictionary_fetch(` we do not see `cache_name: str` but
only `cache_name: TCacheName`.