chore: undo public type aliases (#252)
We use concrete types instead of type aliases for the cache
client and responses. We do this because neither VSCode nor IntelliJ
resolve the type aliases in tooltips when referencing the code; Ie when
typing client.dictionary_fetch( we do not see cache_name: str but
only cache_name: TCacheName. That is too opaque.