feat: make the retry logic configurable
Add abstract EligibilityStrategy class that determines whether a grpc
call can be retried based on the type of call and the error encountered.

Add abstract RetryStrategy class that uses an EligibilityStrategy to
determine if a call can be retried and outputs a time to wait before
retrying.

Add default implementations of the retry and eligibility strategies that
have no wait time.

Update the configurations to take a retry strategy and set the number of
retries to 3 by default. That is, 1 initial call and up to 3 retries on
top of that.