chore: add trace level logging
Do nothing instead of raising an error if a log method is already
defined in the add_logging_level method.

Call initialize_momento_logging in __init__.py to add trace logging.

Add a section to the readme that explains the trace logging and has an
example of how to enable it.