fix: bug where the __repr__ for Cred Provider would mutate it
Prior to this commit, the __repr__ function for a credentials
provider was grabbing a reference to the actual member variables
from the object before trying to obfuscate the token for printing.
Thus, any credential provider whose __repr__ was called would
no longer be valid for use with the cache client from that point
forward.

This commit ensures we do a deep copy of the attributes before
we obfuscate them, leaving the original attributes intact.
Also adds test coverage.