fix: don't call retry logic on OK status codes
This commit changes the retry interceptors so that they don't call
into the retry strategy if the status code is OK.  The main motivation
is that prior to this change we were emitting a debug log message
on every request, even if they were successful.  But generally
speaking there is also no reason to incur any cost overhead for
the retry computations if the request already succeeded.