fix: change library python upper bound to 4
Build templates for python may default to the next python major
version as an upper bound. That is, a project set up for python 3.11
may declare python version compatibility as `^3.11`, which means
`>=3.11,<4.0`. That means when our library declares `>=3.7,<3.12`, we
are potentially incompatible with `^3.11` as this could mean python
versions over `3.11` in the future.