chore: update momento-wire-types, adjust pb/grpc type hints, and refactor requests (#335)
Updates to v0.63.0 of momento-wire-types, which now includes type interfaces. This surfaced a few errors which we fix. Without the type hints the previous implementation seldom used appropriate constructors for requests, so we refactor those. The test directory for the integration tests was out of date ("simple_cache_client") so we update that as well.