fix: prefer `importlib_metadata` to deprecated pkg_resources
Per the
[docs](https://docs.python.org/3.12/library/importlib.metadata.html):

"importlib.metadata is a library that provides access to the metadata
of an installed Distribution Package, such as its entry points or its
top-level names (Import Packages, modules, if any). Built in part on
Python’s import system, this library intends to replace similar
functionality in the entry point API and metadata API of
pkg_resources. Along with importlib.resources, this package can
eliminate the need to use the older and less efficient pkg_resources
package.

importlib.metadata operates on third-party distribution packages
installed into Python’s site-packages directory via tools such as
pip. Specifically, it works with distributions with discoverable
dist-info or egg-info directories, and metadata defined by the Core
metadata specifications."

Prior to this commit, we were using `pkg_resources` to detect the
installed version of momento to include in the agent header. With this
commit we change to using the preferred `importlib.metadata`. Since
That library is standard only in python >= 3.8, we fall back to using
the backported library `importlib_metadata`.