chore: prefer API key over AUTH token terminology (#385)
We have settled on the terminology "API_KEY" instead of
"AUTH_TOKEN". This updates the examples, documentation, and workflows
to use that. An exception are the tests which are internal.

Over the course of this refactor, linting and formatting failed for the examples, so I have corrected errors as well.