feat: allow custom root certificates in cache client (#395)
Like the previous PR for the vector client, this adds the ability to
set the root certificate for the cache client. We add methods to the
top level configuration for convenience. Unit and integration tests
exercise the happy and error paths.

The integration tests for the happy path also exercise when a custom root
certificate is added. Because this reuses a cached certs, we
comment out the tests. They serve as documentation to reproduce good
behavior.

A future PR will add this to the topics client.