feat: add mvi search score threshold
Add the ability to specify a score threshold to prune search results
server side. Search hit distances must meet this threshold to be
included.

For euclidean similarity, scores lower than the threshold are
included; for inner product and cosine similarity, higher scores are
included.

Bumps the protos, implements the feature, and adds tests.