feat: add mvi search and fetch vectors method (#399)
Add MVI search_and_fetch_vectors method and integration tests. This new method performs a search as usual, but returns the vectors associated with the search hits. That is, the only change in behavior comes down to the response.

Because of this, we refactor the search tests to work against both search and search_and_fetch_vectors. We make use of pytest's parameterized test feature to do this.