fix: timedelta to ms conversion and timedelta to deadline conversion (#442)
This PR corrects two bugs.

First we add a utility function to convert timedelta objects to milliseconds, unit tests, and refactor code to use this across the board. Previously we were converting timedeltas to milliseconds in several places which led to bugs when calling .seconds (not fractional) vs total_seconds() (fractional). Calling the former was truncating the milliseconds portion.

Second we correct a bug the gRPC timeout. After reading from the config, we converted the timedelta to total whole seconds. This again truncated the milliseconds portion; because the gRPC parameter accepts fractional seconds, we should not have been truncating