chore: use ubuntu-24.04 for ci where possible (#454)
Updates the CI workflows to use ubuntu-24.04 for python 3.9+ and ubuntu-20.04 for python 3.7 and 3.8. Previously we could not run on ubuntu-22.04 due to a difficult-to-diagnose timeout problem. This timeout problem manifested on Azure VMs with ubuntu-22.04 but not on other platforms.

This problem does not manifest on ubuntu-24.04, so we can use that where possible. Because python 3.7 and python 3.8 are not available on ubuntu-24.04, we still run those on ubuntu-20.04. At some point we can look in to installing via pyenv or manually on ubuntu-24.04.