chore: fall back manual version when dist-info absent (#462)
We temporarily fall back to a manually provided version number when
the dist-info METADATA is absent. This situation occurs in some
frameworks like serverless which strip this for a slim build.

In the near future we will refactor this code so that CI sets the
version number in the file and we do not rely on metadata reflection.