ci: set up with release please v4 and stub publish step
Adds a release please job with v4. Unlike v3, v4 requires a manifest
file for most of the options. We add that to the root of the
project. Unlike v3, we also have to explicitly pass the release please
outputs from the release-please job to the publish job.

We have also verified setting the version in source. That happens with
the extra-files property in the manifest. The version in the main
`__init__` requires a stub version so the release please regex can
substitute it.