chore: add tests for the fixed count retry strategy
Add a middleware that lets tests control Momento Local with metadata and
can collect metrics about retries.

Add sync and async tests for the fixed count retry strategy.

Update the aio retry interceptor to stop if from ignoring the result of
a future in some cases.

Add a make target for local tests.

Add a github action to run the local tests.